canonicalize paths on the command-line as they're used
I'm not entirely happy with my solution.  I think Ninja should
always use canonicalized paths internally, so we should canonicalize
all paths as we receive them (from the command line and from build
files).  But there are other places where we pass paths around
(like in all the tool commands) and I don't want to add manual calls
to canonicalize in each.

Perhaps the longer-term solution is to add some sort of
GetNodeCanonicalize() to State and then make all of the functions that
work with paths as strings (like Builder::AddTarget or RunBrowsePython)
instead only accept a Node*.