make the build log test match the indended build log logic
As described on https://github.com/martine/ninja/pull/349 , this test
shouldn't have been passing.  Fixing a Valgrind issue showed that the
test was wrong.  Thankfully it wasn't a critical issue.

The new code verifies that we either succeed or report an error when
parsing -- the only behavior we really care about is not crashing.