always use output timestamp for deps
Using the timestamp of the .d file was wrong.  It can be written
at a different time than the output, and it is immediately deleted
after parsing; only the output remains for subsequent timestamp
comparison purposes, so always use the output's timestamp.  (This
is how the code worked on Windows already.)