win: Let the "Pause" key or Ctrl-S pause output.
In cmd.exe, hitting the "Pause" key or Ctrl-S will pause
programs until a key is pressed.  This is apparently
implemented when stdout is writing to, so use printf
instead of Console functions to reset the cursor to
the start of the line.  Also happens to simplify the code.

(This already worked in -v mode since that already
prints using printf.)
Based on a patch from gmisocpp@gmail.com!