Remove 'Recompacting...' messages.
Recompacting the build log used to be slow, so it made sense to print this
message.  We then made recompaction much faster, but didn't remove this
message back then.

The deps log only has it because the build log had it.

Since both steps are effectively instant in practice, remove these log messages.