when building with cwd = source root, use relative paths
The pull request that introduced building from a directory other
than the source dir, https://github.com/ninja-build/ninja/pull/979,
made it so all source paths are effectively absolute paths.  This
change restores the old behavior in the case when you are building
in the source.  See the comments there.