support relative ppx paths in .merlin generation
at present, the generation of FLG statements in the .merlin file when a
relative ppx is used contains the string $src_root_dir rather than an
absolute path to the ppx file. this PR changes the
convert_and_resolve_path function to use the provided cwd from parsing
the bsconfig file as the base against which the ppx is resolved.

Signed-Off-By: Jacob Bass <me@bassjacob.com>