[Super errors] Work with -warn-error
There was a part of the warning printing logic that I didn't understand, and temporarily disabled back then because it required a small compiler patch and we didn't decide to do that yet. That line is actually responsible for counting the number of warnings marked as error in -warn-errors, then fatal if it's not 0 at the end.

Fixes #2069