Add support for namespaces.
This works but there's a known bug: if you have a file with the same
name as the package name (say it's A) and you have another file B such
that B depends on A (the file, not the module aliased package-name one),
you'll get a link error `Reference to undefined global 'A'`.
I'm not sure why this is happening.

cc @chenglou @schmavery