So turns out we misunderstood the C linker all this time ...
Turns out the C linker wants the provider of the function first, and
then the user of the function. So:
gcc test2.o test1.o
where test1 depends on test2.
In clang the order doesn't seem to matter, which is why this has been
working on osx, but on linux it seems like the linker will simply drop
the files that aren't "used". Funny.