Add support for "ppx": ["DepName/PpxModuleName"]
Ppx now fully work, only compiled to bytecode. We automatically link
compiler-libs with the ppx also.

Also improved performance by a bunch by not re-crawling the dep tree
n^2 times. Now we only visit each dep once and build the
dependency_info object as we go (the order of the crawling should
guarantee that the record is correct at each level).
There might be a bug where the dependency_info contains the info of
transitive dependencies when linking, which would allow deps' ppx for
example to depend on their transitive dep's C static libraries and such
but eh... That's a tough one to fix (C global namespace and all).

Also add "update" function to String_map, which is the same as the
stdlib update function on map.