Run all ppxes on all source code when building as a dep
This makes the library much slower to compile, but we have no way to
check which file the user is using or which ppx should run on which
file. So we simply run all ppxes on all source files.

This is so expensive. I really need to discourage the use of ppx...